+++ geo-types/src/geometry/polygon.rs
@@ -28,23 +28,23 @@ use approx::{AbsDiffEq, RelativeEq};
 /// # Validity
 ///
 /// - The exterior and interior rings must be valid
-/// `LinearRing`s (see [`LineString`]).
+///   `LinearRing`s (see [`LineString`]).
 ///
 /// - No two rings in the boundary may cross, and may
-/// intersect at a `Point` only as a tangent. In other
-/// words, the rings must be distinct, and for every pair of
-/// common points in two of the rings, there must be a
-/// neighborhood (a topological open set) around one that
-/// does not contain the other point.
+///   intersect at a `Point` only as a tangent. In other
+///   words, the rings must be distinct, and for every pair of
+///   common points in two of the rings, there must be a
+///   neighborhood (a topological open set) around one that
+///   does not contain the other point.
 ///
 /// - The closure of the interior of the `Polygon` must
-/// equal the `Polygon` itself. For instance, the exterior
-/// may not contain a spike.
+///   equal the `Polygon` itself. For instance, the exterior
+///   may not contain a spike.
 ///
 /// - The interior of the polygon must be a connected
-/// point-set. That is, any two distinct points in the
-/// interior must admit a curve between these two that lies
-/// in the interior.
+///   point-set. That is, any two distinct points in the
+///   interior must admit a curve between these two that lies
+///   in the interior.
 ///
 /// Refer to section 6.1.11.1 of the OGC-SFA for a formal
 /// definition of validity. Besides the closed `LineString`