+++ geo/src/algorithm/line_measures/metric_spaces/mod.rs
@@ -1,0 +1,11 @@
+mod euclidean;
+pub use euclidean::Euclidean;
+
+mod geodesic;
+pub use geodesic::Geodesic;
+
+mod haversine;
+pub use haversine::Haversine;
+
+mod rhumb;
+pub use rhumb::Rhumb;