+++ geo/src/algorithm/simplify_vw.rs
@@ -548,9 +548,9 @@ pub trait SimplifyVwPreserve<T, Epsilon = T> {
     /// - It is possible for the simplification algorithm to displace a Polygon's interior ring outside its shell.
     /// - The algorithm does **not** guarantee a valid output geometry, especially on smaller geometries.
     /// - If removal of a point causes a self-intersection, but the geometry only has `n + 1`
-    /// points remaining (3 for a `LineString`, 5 for a `Polygon`), the point is retained and the
-    /// simplification process ends. This is because there is no guarantee that removal of two points will remove
-    /// the intersection, but removal of further points would leave too few points to form a valid geometry.
+    ///   points remaining (3 for a `LineString`, 5 for a `Polygon`), the point is retained and the
+    ///   simplification process ends. This is because there is no guarantee that removal of two points will remove
+    ///   the intersection, but removal of further points would leave too few points to form a valid geometry.
     /// - The tolerance used to remove a point is `epsilon`, in keeping with GEOS. JTS uses `epsilon ^ 2`
     ///
     /// # Examples